Initial UML
Initial plan is to have Pong be the main program with two structs of
type Ball and Player. Pong will have two instances of Player and one
instance of Ball. Initially it would be nice to have multiple max
scores, e.g. best of 1, 3, 5, 7. Can change it to fixed if we have to.