Updated Global Variables and moveBall Function
The moveBall function has been updated to just reflect the ball off of
the top of the screen. Have to expand it to account for bouncing off off
paddles or getting scored. Might have to make the function return a
unsigned char to account for if move was successful or if a goal was
scored.